Remove models.py
Move Match class into scores.py
Move Team class into teams.py
